+++ sdk/metric/meter_test.go
@@ -95,6 +95,8 @@ func TestMeterInstrumentConcurrency(t *testing.T) {
 	wg.Wait()
 }
 
+var emptyCallback metric.Callback = func(ctx context.Context) error { return nil }
+
 // A Meter Should be able register Callbacks Concurrently.
 func TestMeterCallbackCreationConcurrency(t *testing.T) {
 	wg := &sync.WaitGroup{}
@@ -103,11 +105,11 @@ func TestMeterCallbackCreationConcurrency(t *testing.T) {
 	m := NewMeterProvider().Meter("callback-concurrency")
 
 	go func() {
-		_, _ = m.RegisterCallback([]instrument.Asynchronous{}, func(ctx context.Context) {})
+		_, _ = m.RegisterCallback([]instrument.Asynchronous{}, emptyCallback)
 		wg.Done()
 	}()
 	go func() {
-		_, _ = m.RegisterCallback([]instrument.Asynchronous{}, func(ctx context.Context) {})
+		_, _ = m.RegisterCallback([]instrument.Asynchronous{}, emptyCallback)
 		wg.Done()
 	}()
 	wg.Wait()
@@ -115,7 +117,7 @@ func TestMeterCallbackCreationConcurrency(t *testing.T) {
 
 func TestNoopCallbackUnregisterConcurrency(t *testing.T) {
 	m := NewMeterProvider().Meter("noop-unregister-concurrency")
-	reg, err := m.RegisterCallback(nil, func(ctx context.Context) {})
+	reg, err := m.RegisterCallback(nil, emptyCallback)
 	require.NoError(t, err)
 
 	wg := &sync.WaitGroup{}
@@ -143,11 +145,11 @@ func TestCallbackUnregisterConcurrency(t *testing.T) {
 	require.NoError(t, err)
 
 	i := []instrument.Asynchronous{actr}
-	regCtr, err := meter.RegisterCallback(i, func(ctx context.Context) {})
+	regCtr, err := meter.RegisterCallback(i, emptyCallback)
 	require.NoError(t, err)
 
 	i = []instrument.Asynchronous{ag}
-	regG, err := meter.RegisterCallback(i, func(ctx context.Context) {})
+	regG, err := meter.RegisterCallback(i, emptyCallback)
 	require.NoError(t, err)
 
 	wg := &sync.WaitGroup{}
@@ -167,6 +169,7 @@ func TestCallbackUnregisterConcurrency(t *testing.T) {
 
 // Instruments should produce correct ResourceMetrics.
 func TestMeterCreatesInstruments(t *testing.T) {
+	attrs := []attribute.KeyValue{attribute.String("name", "alice")}
 	seven := 7.0
 	testCases := []struct {
 		name string
@@ -176,10 +179,15 @@ func TestMeterCreatesInstruments(t *testing.T) {
 		{
 			name: "ObservableInt64Count",
 			fn: func(t *testing.T, m metric.Meter) {
-				ctr, err := m.Int64ObservableCounter("aint")
+				cback := func(ctx context.Context, o instrument.Int64Observer) error {
+					o.Observe(ctx, 4, attrs...)
+					return nil
+				}
+				ctr, err := m.Int64ObservableCounter("aint", instrument.WithInt64Callback(cback))
 				assert.NoError(t, err)
-				_, err = m.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) {
+				_, err = m.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) error {
 					ctr.Observe(ctx, 3)
+					return nil
 				})
 				assert.NoError(t, err)
 
@@ -192,6 +200,7 @@ func TestMeterCreatesInstruments(t *testing.T) {
 					Temporality: metricdata.CumulativeTemporality,
 					IsMonotonic: true,
 					DataPoints: []metricdata.DataPoint[int64]{
+						{Attributes: attribute.NewSet(attrs...), Value: 4},
 						{Value: 3},
 					},
 				},
@@ -200,10 +209,15 @@ func TestMeterCreatesInstruments(t *testing.T) {
 		{
 			name: "ObservableInt64UpDownCount",
 			fn: func(t *testing.T, m metric.Meter) {
-				ctr, err := m.Int64ObservableUpDownCounter("aint")
+				cback := func(ctx context.Context, o instrument.Int64Observer) error {
+					o.Observe(ctx, 4, attrs...)
+					return nil
+				}
+				ctr, err := m.Int64ObservableUpDownCounter("aint", instrument.WithInt64Callback(cback))
 				assert.NoError(t, err)
-				_, err = m.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) {
+				_, err = m.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) error {
 					ctr.Observe(ctx, 11)
+					return nil
 				})
 				assert.NoError(t, err)
 
@@ -216,6 +230,7 @@ func TestMeterCreatesInstruments(t *testing.T) {
 					Temporality: metricdata.CumulativeTemporality,
 					IsMonotonic: false,
 					DataPoints: []metricdata.DataPoint[int64]{
+						{Attributes: attribute.NewSet(attrs...), Value: 4},
 						{Value: 11},
 					},
 				},
@@ -224,10 +239,15 @@ func TestMeterCreatesInstruments(t *testing.T) {
 		{
 			name: "ObservableInt64Gauge",
 			fn: func(t *testing.T, m metric.Meter) {
-				gauge, err := m.Int64ObservableGauge("agauge")
+				cback := func(ctx context.Context, o instrument.Int64Observer) error {
+					o.Observe(ctx, 4, attrs...)
+					return nil
+				}
+				gauge, err := m.Int64ObservableGauge("agauge", instrument.WithInt64Callback(cback))
 				assert.NoError(t, err)
-				_, err = m.RegisterCallback([]instrument.Asynchronous{gauge}, func(ctx context.Context) {
+				_, err = m.RegisterCallback([]instrument.Asynchronous{gauge}, func(ctx context.Context) error {
 					gauge.Observe(ctx, 11)
+					return nil
 				})
 				assert.NoError(t, err)
 
@@ -238,6 +258,7 @@ func TestMeterCreatesInstruments(t *testing.T) {
 				Name: "agauge",
 				Data: metricdata.Gauge[int64]{
 					DataPoints: []metricdata.DataPoint[int64]{
+						{Attributes: attribute.NewSet(attrs...), Value: 4},
 						{Value: 11},
 					},
 				},
@@ -246,10 +267,15 @@ func TestMeterCreatesInstruments(t *testing.T) {
 		{
 			name: "ObservableFloat64Count",
 			fn: func(t *testing.T, m metric.Meter) {
-				ctr, err := m.Float64ObservableCounter("afloat")
+				cback := func(ctx context.Context, o instrument.Float64Observer) error {
+					o.Observe(ctx, 4, attrs...)
+					return nil
+				}
+				ctr, err := m.Float64ObservableCounter("afloat", instrument.WithFloat64Callback(cback))
 				assert.NoError(t, err)
-				_, err = m.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) {
+				_, err = m.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) error {
 					ctr.Observe(ctx, 3)
+					return nil
 				})
 				assert.NoError(t, err)
 
@@ -262,6 +288,7 @@ func TestMeterCreatesInstruments(t *testing.T) {
 					Temporality: metricdata.CumulativeTemporality,
 					IsMonotonic: true,
 					DataPoints: []metricdata.DataPoint[float64]{
+						{Attributes: attribute.NewSet(attrs...), Value: 4},
 						{Value: 3},
 					},
 				},
@@ -270,10 +297,15 @@ func TestMeterCreatesInstruments(t *testing.T) {
 		{
 			name: "ObservableFloat64UpDownCount",
 			fn: func(t *testing.T, m metric.Meter) {
-				ctr, err := m.Float64ObservableUpDownCounter("afloat")
+				cback := func(ctx context.Context, o instrument.Float64Observer) error {
+					o.Observe(ctx, 4, attrs...)
+					return nil
+				}
+				ctr, err := m.Float64ObservableUpDownCounter("afloat", instrument.WithFloat64Callback(cback))
 				assert.NoError(t, err)
-				_, err = m.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) {
+				_, err = m.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) error {
 					ctr.Observe(ctx, 11)
+					return nil
 				})
 				assert.NoError(t, err)
 
@@ -286,6 +318,7 @@ func TestMeterCreatesInstruments(t *testing.T) {
 					Temporality: metricdata.CumulativeTemporality,
 					IsMonotonic: false,
 					DataPoints: []metricdata.DataPoint[float64]{
+						{Attributes: attribute.NewSet(attrs...), Value: 4},
 						{Value: 11},
 					},
 				},
@@ -294,10 +327,15 @@ func TestMeterCreatesInstruments(t *testing.T) {
 		{
 			name: "ObservableFloat64Gauge",
 			fn: func(t *testing.T, m metric.Meter) {
-				gauge, err := m.Float64ObservableGauge("agauge")
+				cback := func(ctx context.Context, o instrument.Float64Observer) error {
+					o.Observe(ctx, 4, attrs...)
+					return nil
+				}
+				gauge, err := m.Float64ObservableGauge("agauge", instrument.WithFloat64Callback(cback))
 				assert.NoError(t, err)
-				_, err = m.RegisterCallback([]instrument.Asynchronous{gauge}, func(ctx context.Context) {
+				_, err = m.RegisterCallback([]instrument.Asynchronous{gauge}, func(ctx context.Context) error {
 					gauge.Observe(ctx, 11)
+					return nil
 				})
 				assert.NoError(t, err)
 
@@ -308,6 +346,7 @@ func TestMeterCreatesInstruments(t *testing.T) {
 				Name: "agauge",
 				Data: metricdata.Gauge[float64]{
 					DataPoints: []metricdata.DataPoint[float64]{
+						{Attributes: attribute.NewSet(attrs...), Value: 4},
 						{Value: 11},
 					},
 				},
@@ -470,16 +509,18 @@ func TestMetersProvideScope(t *testing.T) {
 	m1 := mp.Meter("scope1")
 	ctr1, err := m1.Float64ObservableCounter("ctr1")
 	assert.NoError(t, err)
-	_, err = m1.RegisterCallback([]instrument.Asynchronous{ctr1}, func(ctx context.Context) {
+	_, err = m1.RegisterCallback([]instrument.Asynchronous{ctr1}, func(ctx context.Context) error {
 		ctr1.Observe(ctx, 5)
+		return nil
 	})
 	assert.NoError(t, err)
 
 	m2 := mp.Meter("scope2")
 	ctr2, err := m2.Int64ObservableCounter("ctr2")
 	assert.NoError(t, err)
-	_, err = m1.RegisterCallback([]instrument.Asynchronous{ctr2}, func(ctx context.Context) {
+	_, err = m1.RegisterCallback([]instrument.Asynchronous{ctr2}, func(ctx context.Context) error {
 		ctr2.Observe(ctx, 7)
+		return nil
 	})
 	assert.NoError(t, err)
 
@@ -563,7 +604,10 @@ func TestUnregisterUnregisters(t *testing.T) {
 		floag64Counter,
 		floag64UpDownCounter,
 		floag64Gauge,
-	}, func(context.Context) { called = true })
+	}, func(context.Context) error {
+		called = true
+		return nil
+	})
 	require.NoError(t, err)
 
 	ctx := context.Background()
@@ -613,7 +657,10 @@ func TestRegisterCallbackDropAggregations(t *testing.T) {
 		floag64Counter,
 		floag64UpDownCounter,
 		floag64Gauge,
-	}, func(context.Context) { called = true })
+	}, func(context.Context) error {
+		called = true
+		return nil
+	})
 	require.NoError(t, err)
 
 	data, err := r.Collect(context.Background())
@@ -643,10 +690,11 @@ func testAttributeFilter(temporality metricdata.Temporality) func(*testing.T) {
 				if err != nil {
 					return err
 				}
-				_, err = mtr.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) {
+				_, err = mtr.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) error {
 					ctr.Observe(ctx, 1.0, attribute.String("foo", "bar"), attribute.Int("version", 1))
 					ctr.Observe(ctx, 2.0, attribute.String("foo", "bar"))
 					ctr.Observe(ctx, 1.0, attribute.String("foo", "bar"), attribute.Int("version", 2))
+					return nil
 				})
 				return err
 			},
@@ -671,10 +719,11 @@ func testAttributeFilter(temporality metricdata.Temporality) func(*testing.T) {
 				if err != nil {
 					return err
 				}
-				_, err = mtr.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) {
+				_, err = mtr.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) error {
 					ctr.Observe(ctx, 1.0, attribute.String("foo", "bar"), attribute.Int("version", 1))
 					ctr.Observe(ctx, 2.0, attribute.String("foo", "bar"))
 					ctr.Observe(ctx, 1.0, attribute.String("foo", "bar"), attribute.Int("version", 2))
+					return nil
 				})
 				return err
 			},
@@ -699,9 +748,10 @@ func testAttributeFilter(temporality metricdata.Temporality) func(*testing.T) {
 				if err != nil {
 					return err
 				}
-				_, err = mtr.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) {
+				_, err = mtr.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) error {
 					ctr.Observe(ctx, 1.0, attribute.String("foo", "bar"), attribute.Int("version", 1))
 					ctr.Observe(ctx, 2.0, attribute.String("foo", "bar"), attribute.Int("version", 2))
+					return nil
 				})
 				return err
 			},
@@ -724,10 +774,11 @@ func testAttributeFilter(temporality metricdata.Temporality) func(*testing.T) {
 				if err != nil {
 					return err
 				}
-				_, err = mtr.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) {
+				_, err = mtr.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) error {
 					ctr.Observe(ctx, 10, attribute.String("foo", "bar"), attribute.Int("version", 1))
 					ctr.Observe(ctx, 20, attribute.String("foo", "bar"))
 					ctr.Observe(ctx, 10, attribute.String("foo", "bar"), attribute.Int("version", 2))
+					return nil
 				})
 				return err
 			},
@@ -752,10 +803,11 @@ func testAttributeFilter(temporality metricdata.Temporality) func(*testing.T) {
 				if err != nil {
 					return err
 				}
-				_, err = mtr.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) {
+				_, err = mtr.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) error {
 					ctr.Observe(ctx, 10, attribute.String("foo", "bar"), attribute.Int("version", 1))
 					ctr.Observe(ctx, 20, attribute.String("foo", "bar"))
 					ctr.Observe(ctx, 10, attribute.String("foo", "bar"), attribute.Int("version", 2))
+					return nil
 				})
 				return err
 			},
@@ -780,9 +832,10 @@ func testAttributeFilter(temporality metricdata.Temporality) func(*testing.T) {
 				if err != nil {
 					return err
 				}
-				_, err = mtr.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) {
+				_, err = mtr.RegisterCallback([]instrument.Asynchronous{ctr}, func(ctx context.Context) error {
 					ctr.Observe(ctx, 10, attribute.String("foo", "bar"), attribute.Int("version", 1))
 					ctr.Observe(ctx, 20, attribute.String("foo", "bar"), attribute.Int("version", 2))
+					return nil
 				})
 				return err
 			},